+++ src/main/java/ru/macdi/demo/repository/UserDAOImpl.java
@@ -29,12 +29,12 @@ public class UserDAOImpl implements UserDAO {
     }
 
     @Override
-    public User getUserById(int id) {
+    public User getUserById(Long id) {
         return entityManager.find(User.class, id);
     }
 
     @Override
-    public void deleteUser(int id) {
+    public void deleteUser(Long id) {
         entityManager.createQuery("delete from User where id=: id")
                 .setParameter("id", id).executeUpdate();
     }
